LayoutTests/http/wpt/service-workers/fetch-service-worker-preload-changing-request.https.html

<!doctype html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/routines.js"></script>
</head>
<body>
<script>
const url = "/WebKit/service-workers/resources/fetch-service-worker-preload-script.py?getResponseFromRequestWithCustomHeader=yes&customHeader=1&token=" + token();

promise_test(async (test) => {
    return setupActivatedButNotRunningServiceWorker("/WebKit/service-workers/fetch-service-worker-preload-worker.js", url);
}, "Setup activated but not running worker");

promise_test(async (test) => {
    await fetch(url + "&value=use-preload", { method: 'POST' });

    // Load iframe, with not running worker, so only preload will start.
    const promise = withIframe(url);

    const frame = await promise;
    assert_equals(frame.contentWindow.value, "my-custom-header");
}, "Service worker load does not use preload when using an updated fetch request");
</script>
</body>
</html>
